+++ src/Blazar/Blazar/Renderer/Buffer.h
@@ -95,7 +95,7 @@ class IndexBuffer {
     virtual ~IndexBuffer() {}
     virtual void Bind() const = 0;
     virtual void Unbind() const = 0;
-    virtual uint32_t GetCount() const = 0;
+    virtual size_t GetCount() const = 0;
     // Factory method
     static Ref<IndexBuffer> Create(uint32_t* verticies, size_t size);
 };
